+++ src/Spectre.Cli/Command`1.cs
@@ -1,4 +1,4 @@
-﻿using System.Diagnostics;
+using System.Diagnostics;
 using System.Threading.Tasks;
 
 namespace Spectre.Cli
@@ -26,7 +26,7 @@ namespace Spectre.Cli
         /// </summary>
         /// <param name="context">The command context.</param>
         /// <param name="settings">The settings.</param>
-        /// <returns>The validation result.</returns>
+        /// <returns>An integer indicating whether or not the command executed successfully.</returns>
         public abstract int Execute(CommandContext context, TSettings settings);
 
         ValidationResult ICommand.Validate(CommandContext context, object settings)